V5.7.1

+++ src/XrdTls/XrdTlsContext.cc
@@ -591,7 +591,7 @@ XrdTlsContext::XrdTlsContext(const char *cert,  const char *key,
          SSL_CTX **ctxLoc;
         } ctx_tracker(&pImpl->ctx);
 
-   static const int sslOpts = SSL_OP_ALL
+   static const uint64_t sslOpts = SSL_OP_ALL
                             | SSL_OP_NO_SSLv2
                             | SSL_OP_NO_SSLv3
                             | SSL_OP_NO_COMPRESSION
